Working with states moved to CommandExecutor
