+++ src/main/java/com/odeyalo/bot/suiri/service/command/steps/OriginalWordAddNewWordMessageBuildingStep.java
@@ -1,8 +1,8 @@
 package com.odeyalo.bot.suiri.service.command.steps;
 
 import com.odeyalo.bot.suiri.domain.AddNewWordMessage;
-import com.odeyalo.bot.suiri.service.command.support.AddNewWordState;
-import com.odeyalo.bot.suiri.service.command.support.AddNewWordStateRepository;
+import com.odeyalo.bot.suiri.service.command.support.state.AddNewWordState;
+import com.odeyalo.bot.suiri.service.command.support.state.AddNewWordStateRepository;
 import com.odeyalo.bot.suiri.support.TelegramUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -29,11 +29,9 @@ public class OriginalWordAddNewWordMessageBuildingStep extends AbstractAddNewWor
             this.logger.warn("The step was skipped since state was wrong. Expected: {}, received: {}",  currentState, getState());
             return null;
         }
-
         String text = TelegramUtils.getText(update);
         this.logger.info("Received original word: {}", text);
         message.setOriginalWord(text);
-        this.stateRepository.saveState(chatId, AddNewWordState.TRANSLATED_WORD);
         return new SendMessage(chatId, "Now send the translated word");
     }
