+++ src/main/java/com/odeyalo/bot/suiri/service/command/steps/TranslatedWordAddNewWordMessageBuildingStep.java
@@ -1,8 +1,8 @@
 package com.odeyalo.bot.suiri.service.command.steps;
 
 import com.odeyalo.bot.suiri.domain.AddNewWordMessage;
-import com.odeyalo.bot.suiri.service.command.support.AddNewWordState;
-import com.odeyalo.bot.suiri.service.command.support.AddNewWordStateRepository;
+import com.odeyalo.bot.suiri.service.command.support.state.AddNewWordState;
+import com.odeyalo.bot.suiri.service.command.support.state.AddNewWordStateRepository;
 import com.odeyalo.bot.suiri.support.TelegramUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -34,7 +34,6 @@ public class TranslatedWordAddNewWordMessageBuildingStep extends AbstractAddNewW
         String text = TelegramUtils.getText(update);
         this.logger.info("Received translated word: {}", text);
         message.setTranslatedWords(Collections.singletonList(text));
-        this.stateRepository.saveState(chatId, AddNewWordState.PICTURE);
         return new SendMessage(chatId, "Now send the picture");
     }
